pyproject.toml
[tool.poetry]
name = "langchain-demo"
version = "0.1.0"
description = ""
authors = ["Ben Van Sleen <benvansleen@gmail.com>"]

[tool.poetry.dependencies]
python = ">3.9.7,<4.0"
langchain = "^0.0.132"
streamlit = "^1.20.0"
python-dotenv = "^1.0.0"
openai = "^0.27.4"
huggingface-hub = "^0.13.3"
google-search-results = "^2.4.2"

[tool.poetry.dev-dependencies]
autopep8 = "^2.0.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.autopep8]
max_line_length = 80
aggressive = 3
in_place = true
recursive = true